pyproject.toml
[project]
name = "quantmini"
version = "0.1.0"
description = "High-Performance Data Pipeline for Financial Market Data"
requires-python = ">=3.10"
dependencies = [
    "pyarrow>=18.1.0",
    "polars>=1.18.0",
    "pandas>=2.2.3",
    "aioboto3>=13.0.1",
    "boto3>=1.35.74",
    "pyyaml>=6.0.2",
    "psutil>=6.1.1",
    "duckdb>=1.0.0",
    "pyqlib>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "e2e: End-to-end integration tests",
    "slow: Slow-running tests",
]
